requirements.txt
# Data
numpy>=1.21
pandas>=1.3

# Visualisation
matplotlib>=3.4
seaborn>=0.11
plotly>=5.0

# Imagerie
opencv-python>=4.5
scikit-image>=0.18
Pillow>=8.2
albumentations>=1.0

# ML classique
scikit-learn>=1.0
joblib>=1.0
optuna>=2.0

# Deep Learning (PyTorch)
torch>=2.0
torchvision>=0.15

# Notebook
jupyterlab>=3.0
notebook>=6.0

# Web Framework
flask>=2.0